zero compiler warnings

+++ TerrainOperations/dllmain.cpp
@@ -32,7 +32,7 @@ ex double terrainops_generate(float* data, float* out, double width, double heig
 }
 
 ex double terrainops_build_settings(double save_all, double swap_zup, double swap_uv, double center, double density, double width, double height, double scale) {
-	terrainops::build_settings((bool)save_all, (bool)swap_zup, (bool)swap_uv, (bool)center, (int)density, (float)width, (float)height, (float)scale);
+	terrainops::build_settings((bool)save_all, (bool)swap_zup, (bool)swap_uv, (bool)center, (int)density, (int)width, (int)height, (float)scale);
 	return 1.0;
 }
