+++ TerrainOperations/terrainops.h
@@ -10,8 +10,8 @@ namespace terrainops {
 	extern bool save_swap_uv;
 	extern bool save_centered;
 	extern int save_density;
-	extern float save_width;
-	extern float save_height;
+	extern int save_width;
+	extern int save_height;
 	extern float save_scale;
 
 	const char* version();
@@ -20,12 +20,12 @@ namespace terrainops {
 	void from_heightmap(float*, unsigned int*, int, float);
 	void flatten(float*, float*, int, float);
 	void apply_scale(float*, float*, int, float);
-	void build_settings(bool, bool, bool, bool, int, float, float, float);
+	void build_settings(bool, bool, bool, bool, int, int, int, float);
 	long build(float*, long long*, int, int);
 	void generate(float*, float*, int, int);
 
 	float get_z(float*, int, int, int h);
-	inline void write_vertex(float*, int*, float, float, float, float, float, unsigned int, float, float, float);
+	inline void write_vertex(float*, long long*, float, float, float, float, float, unsigned int, float, float, float);
 }
 
 #endif
